+++ src/py42/services/alertrules.py
@@ -1,7 +1,5 @@
-from py42.exceptions import Py42BadRequestError
 from py42.exceptions import Py42InvalidRuleError
 from py42.exceptions import Py42NotFoundError
-from py42.exceptions import Py42UserNotOnListError
 from py42.services import BaseService
 
 
@@ -43,11 +41,7 @@ class AlertRulesService(BaseService):
 
     def add_user(self, rule_id, user_id):
         tenant_id = self._user_context.get_current_tenant_id()
-        try:
-            self._user_profile_service.create_if_not_exists(user_id)
-            user_details = self._user_profile_service.get_by_id(user_id)
-        except (Py42BadRequestError, Py42NotFoundError) as err:
-            raise Py42UserNotOnListError(err, user_id, "user profile")
+        user_details = self._user_profile_service.get_by_id(user_id)
         user_aliases = user_details.data.get(u"cloudUsernames") or []
         data = {
             u"tenantId": tenant_id,