+++ src/py42/services/legalhold.py
@@ -7,6 +7,7 @@ from py42.exceptions import Py42LegalHoldNotFoundOrPermissionDeniedError
 from py42.exceptions import Py42UserAlreadyAddedError
 from py42.services import BaseService
 from py42.services.util import get_all_pages
+from py42.util import parse_timestamp_to_milliseconds_precision
 
 
 def _active_state_map(active):
@@ -273,6 +274,75 @@ class LegalHoldService(BaseService):
             user_uid=user_uid,
             user=user,
             active=active,
+        )
+
+    def get_events_page(
+        self,
+        legal_hold_uid=None,
+        min_event_date=None,
+        max_event_date=None,
+        page_num=1,
+        page_size=None,
+    ):
+        """Gets an individual page of Legal Hold events.
+
+        `REST Documentation <https://console.us.code42.com/apidocviewer/#LegalHoldEventReport-get>`__
+
+        Args:
+            legal_hold_uid (str, optional): Find LegalHoldEvents for the Legal Hold Matter
+                with this unique identifier. Defaults to None.
+            min_event_date (ISO date/time string, optional): Find LegalHoldEvents whose
+                eventDate is equal to or after this time. Defaults to None.
+            max_event_date (ISO date/time string, optional): Find LegalHoldEvents whose
+                eventDate is equal to or before this time. Defaults to None.
+            page_num (int): The page number to request. Defaults to 1.
+            page_size (int, optional): The size of the page. Defaults to `py42.settings.items_per_page`.
+
+        Returns:
+            :class:`py42.response.Py42Response`:
+        """
+        page_size = page_size or settings.items_per_page
+        if min_event_date:
+            min_event_date = parse_timestamp_to_milliseconds_precision(min_event_date)
+        if max_event_date:
+            max_event_date = parse_timestamp_to_milliseconds_precision(max_event_date)
+        params = {
+            u"legalHoldUid": legal_hold_uid,
+            u"minEventDate": min_event_date,
+            u"maxEventDate": max_event_date,
+            u"pgNum": page_num,
+            u"pgSize": page_size,
+        }
+        uri = u"/api/LegalHoldEventReport"
+
+        return self._connection.get(uri, params=params)
+
+    def get_all_events(
+        self, legal_hold_uid=None, min_event_date=None, max_event_date=None
+    ):
+        """Gets an individual page of Legal Hold events.
+        `REST Documentation <https://console.us.code42.com/apidocviewer/#LegalHoldEventReport-get>`__
+
+        Args:
+            legal_hold_uid (str, optional): Find LegalHoldEvents for the Legal Hold Matter
+                with this unique identifier. Defaults to None.
+            min_event_date (date/time string, optional): Find LegalHoldEvents whose
+                eventDate is equal to or after this time. E.g. yyyy-MM-dd HH:MM:SS
+                Defaults to None.
+            max_event_date (date/time string, optional): Find LegalHoldEvents whose
+                eventDate is equal to or before this time. E.g. yyyy-MM-dd HH:MM:SS
+                Defaults to None.
+
+        Returns:
+            generator: An object that iterates over :class:`py42.response.Py42Response` objects
+            that each contain a page of LegalHoldEvent objects.
+        """
+        return get_all_pages(
+            self.get_events_page,
+            u"legalHoldEvents",
+            legal_hold_uid=legal_hold_uid,
+            min_event_date=min_event_date,
+            max_event_date=max_event_date,
         )
 
     def add_to_matter(self, user_uid, legal_hold_uid):